# src/APIGateway/APIGateway.API/wwwroot/openapi — копия.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Memories – Unified API",
    "version": "1.1.1",
    "description": "Gateway spec combining microservices: Identity, MemoryArchive, Profile, Notification, AccessControl, AuditLogging"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Gateway"
    }
  ],
...
  "security": [
    { "Bearer": [] }
  ],
  "paths": {
...
    "/identity/register": {
      "post": {
        "tags": [ "Identity" ],
        "summary": "Create a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Created (returns JWT)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JwtResponse" } } }
          },
          "409": { "description": "User already exists" }
        }
      }
    },
    "/identity/login": {
      "post": {
        "tags": [ "Identity" ],
        "summary": "User login (returns JWT)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JwtResponse" } } }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/identity/refresh": {
      "post": {
        "tags": [ "Identity" ],
        "summary": "Refresh JWT",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JwtResponse" } } }
          }
        }
      }
    },
    "/identity/me": {
      "get": {
        "tags": [ "Identity" ],
        "summary": "Get current user profile (from token)",
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/access-control/api/access/grant": {
      "post": {
        "tags": [ "AccessControl" ],
        "summary": "Grant access",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GrantAccessRequest" } } }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AccessRuleDto" } } }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/access-control/api/access/revoke": {
      "post": {
        "tags": [ "AccessControl" ],
        "summary": "Revoke access",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccessRevokeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Revoked (idempotent)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "removed": { "type": "integer" },
                    "subjectId": { "type": "string" },
                    "subjectType": { "type": "string" },
                    "resourceType": { "type": "string" },
                    "resourceId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "permission": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/access-control/api/access/check": {
      "post": {
        "tags": [ "AccessControl" ],
        "summary": "Check access",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckAccessRequest" } } }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckAccessResponse" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/access-control/api/access/memory/{id}/shares": {
      "get": {
        "tags": [ "AccessControl" ],
        "summary": "List shares for a Memory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AccessRuleDto" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide JWT as: Bearer {token}"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": [ "username", "password" ],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [ "username", "password", "email" ],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" },
          "email": {
            "type": "string",
            "format": "email"
          },
          "displayName": { "type": "string" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": [ "refreshToken" ],
        "properties": { "refreshToken": { "type": "string" } }
      },
      "JwtResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GrantAccessRequest": {
        "type": "object",
        "required": [ "subjectId", "subjectType", "resourceType", "resourceId", "permission" ],
        "properties": {
          "subjectId": {
            "type": "string",
            "description": "UUID пользователя или имя роли"
          },
          "subjectType": {
            "type": "string",
            "enum": [ "User", "Role" ]
          },
          "resourceType": {
            "type": "string",
            "enum": [ "Memory" ]
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "permission": {
            "type": "string",
            "enum": [ "View", "Edit", "Delete" ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "CheckAccessRequest": {
        "type": "object",
        "required": [ "subjectId", "resourceType", "resourceId", "permission" ],
        "properties": {
          "subjectId": { "type": "string" },
          "resourceType": {
            "type": "string",
            "enum": [ "Memory" ]
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "permission": {
            "type": "string",
            "enum": [ "View", "Edit", "Delete" ]
          }
        }
      },
      "AccessRuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": { "type": "string" },
          "subjectType": { "type": "string" },
          "resourceType": { "type": "string" },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "permission": { "type": "string" },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CheckAccessResponse": {
        "type": "object",
        "properties": { "allowed": { "type": "boolean" } }
      }
    }
  }
}